+++ apps/web/app/departments/[id]/task_add/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import TaskAdd from "@/components/tasks/form/task.add";
+import TaskAdd from "@/components/departments/task_form/task.add";
 import dynamic from "next/dynamic";
 import { useRouter } from "next/navigation";
