+++ apps/web/components/tasks/list/index.tsx
@@ -3,7 +3,6 @@
 import { DataTablePagination } from "@/components/common/list/list.pagination";
 import LoaderSkeleton from "@/components/common/list/loder.skeleton";
 import { useClosedTask, useOpenTask } from "@/hooks/subgraph/task";
-import { PATHS } from "@/routes/paths";
 import hasRole from "@/utils/role";
 import {
   ColumnFiltersState,
@@ -18,14 +17,12 @@ import {
   VisibilityState,
 } from "@tanstack/react-table";
 import { TaskCreated } from "@workspace/sdk/types/task.type";
-import { Button } from "@workspace/ui/components/button";
 import {
   Tabs,
   TabsContent,
   TabsList,
   TabsTrigger,
 } from "@workspace/ui/components/tabs";
-import { Plus } from "lucide-react";
 import { AppRouterInstance } from "next/dist/shared/lib/app-router-context.shared-runtime";
 import { useState } from "react";
 import { useColumns } from "../details/details.column";
@@ -151,39 +148,6 @@ export default function TaskListMain({ router }: TaskListMainProps) {
             <h3 className="text-gray-500 font-normal text-sm">
               List of all the tasks
             </h3>
-          </div>
-
-          <div className="flex flex-col ml-auto justify-end h-full space-y-2">
-            {/* Warning message always visible */}
-            <div className="flex items-center gap-2 bg-yellow-100 border-l-4 border-yellow-500 text-yellow-800 p-3 rounded-md shadow-sm max-w-xs">
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                className="h-5 w-5 flex-shrink-0"
-                fill="none"
-                viewBox="0 0 24 24"
-                stroke="currentColor"
-                strokeWidth={2.5}
-              >
-                <path
-                  strokeLinecap="round"
-                  strokeLinejoin="round"
-                  d="M12 9v2m0 4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
-                />
-              </svg>
-              <span className="text-sm font-medium">
-                Only entity owners can create tasks.
-              </span>
-            </div>
-
-            {/* Create Task Button */}
-            <Button
-              className="min-w-[10rem] font-semibold h-10"
-              variant="default"
-              onClick={() => router.push(PATHS.TASKS.ADD)}
-            >
-              <Plus size={22} strokeWidth={2.75} />
-              <span>Create Task</span>
-            </Button>
           </div>
         </div>
